+++ packages/reactor-local/src/util.ts
@@ -74,15 +74,34 @@ export async function addDefaultDrive(
 
 export async function startViteServer() {
   const vite = await createServer({
-    server: { middlewareMode: true, watch: null },
+    server: { middlewareMode: true },
     appType: "custom",
     build: {
       rollupOptions: {
         input: [],
       },
     },
-    plugins: [viteCommonjs()],
+    plugins: [
+      viteCommonjs(),
+      {
+        name: "suppress-hmr",
+        handleHotUpdate() {
+          return []; // return empty array to suppress server refresh
+        },
+      },
+    ],
   });
 
   return vite;
+}
+
+export function debounce<T extends (...args: any[]) => void>(
+  func: T,
+  delay = 100,
+): (...args: Parameters<T>) => void {
+  let timeoutId: ReturnType<typeof setTimeout>;
+  return (...args: Parameters<T>): void => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => func(...args), delay);
+  };
 }